Relax restrictions when using the foreign implementations
The foreign implementations don't require the input array dimensions to be powers of two, so push the check down into the default implementation only. This also means that the Haskell-side size parameter is ignored by the foreign implementations, which can make the API easier to use (this is somewhat unsafe, but requiring the FFI implementations else error is reasonable).