Improve FFT*Many_plans functions
FFT2Many_plans created a plan for usual 2D FFT
  (that is why fft and fft2D return the same results
   with PTX backend and different results with Native backend).
The same problem with FFT3Many_plans function.

I've changed the "layout" to a list with single Int (length of the innermost dimension)
and the "batched" to a number of 1D FFTs (h in case of a 2D array and d*h in case of a 3D array)

Now, 1D FFTs over 2- and 3-dimensional arrays are performed correctly.